+++ CHANGELOG.md
@@ -9,5 +9,8 @@
   * (breaking change) `preload` reducer is now added by default and "preload" reducer name is now reserved for it. `<Loading/>` component is also now built-in into the library and can be `import`ed from it.
   * (breaking change) Removed `request` from rendering service `initialize()` parameters (may be added back, upon request).
   * (breaking change) Removed `loading` from rendering service parameters (seems that it was never used).
-  * (breaking change) `render: false` parameter renamed to `hollow: true`
-  * `reduxEventNaming` and `reduxPropertyNaming` are now set by default
+  * (breaking change) `render: false` parameter renamed to `hollow: true`.
+  * `reduxEventNaming` and `reduxPropertyNaming` are now set by default.
+  * (breaking change) `@preload()` now doesn't automatically convert `Array`s into `Promise.all(array)`.
+  * (breaking change) `@preload()` `helpers` removed (may be added back upon request).
+  * (breaking change) Removed `{ preload: { client } }` configuration parameter.